refactor: extract keymap methods into separate files
